@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use App\Models\Product;
-use App\Services\Filters\Filter;
+use App\Services\Filters\PlainFilter;
 use App\Services\Filters\SliderFilter;
 
 class SearchController extends Controller
@@ -37,7 +37,7 @@ class SearchController extends Controller
             'price' => SliderFilter::init('products.price', 'price', 'Price')
                 ->fetchValues($query)
                 ->fetchParams($request->filter),
-            'brand' => Filter::init('products.brand', 'brand', 'Brand')
+            'brand' => PlainFilter::init('products.brand', 'brand', 'Brand')
                 ->fetchValues($query)
                 ->fetchParams($request->filter),
         ])->each(function ($element) use ($query) {
