@@ -144,7 +144,7 @@
                     })
                     .catch(e => {
                         this.isLoading = false
-                        this.$buefy.toast.open({
+                        this.$buefy.toasts.open({
                             message: `Error: ${e.message}`,
                             type: 'is-danger'
                         })
