@@ -1,10 +1,10 @@
 <div class="p-3 text-center" id="info-btn-group">
-    <button class="btn navbar-icon-btn icon-30 icon-phone" data-bs-parent="#info-btn-group"
+    <button class="btn navbar-icon-btn icon-30 icon-phone" data-bs-parent="#navbar-info"
             data-bs-target="#navbar-info-phone" data-bs-toggle="collapse"></button>
-    <button class="btn navbar-icon-btn icon-30 icon-search" data-bs-parent="#info-btn-group"
+    <button class="btn navbar-icon-btn icon-30 icon-search" data-bs-parent="#navbar-info"
             data-bs-target="#navbar-info-search" data-bs-toggle="collapse"></button>
-    <button class="btn navbar-icon-btn icon-30 icon-compare" data-bs-parent="#info-btn-group"
-            data-bs-target="#navbar-info-compare" data-bs-toggle="collapse"></button>
+{{--    <button class="btn navbar-icon-btn icon-30 icon-compare" data-bs-parent="#navbar-info"--}}
+{{--            data-bs-target="#navbar-info-compare" data-bs-toggle="collapse"></button>--}}
     <button id="cart-status" class="btn navbar-icon-btn icon-30 icon-cart-empty"
-            data-bs-target="#modal-cart" data-bs-toggle="modal"></button>
+            data-bs-target="#modal-shopping-cart" data-bs-toggle="modal"></button>
 </div>
