@@ -1,12 +1,12 @@
-<div id="modal-cart" class="modal fade">
-    <div class="modal-dialog">
+<div class="modal" id="modal-shopping-cart" tabIndex="-1" aria-hidden="true">
+    <div class="modal-dialog modal-fullscreen modal-dialog-scrollable">
         <div class="modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal">&times;</button>
+            <div class="modal-header px-4">
+                <h3 class="modal-title" id="exampleModalLabel">Корзина покупок</h3>
+                <button type="button" class="btn-close" data-bs-dismiss="modal"/>
             </div>
-            <h4><span>Корзина покупок</span></h4>
             <div class="modal-body text-center">
-
+                <div id="shopping-cart" />
             </div>
         </div>
     </div>
